requirements.txt
asgiref==3.7.2
attrs==23.1.0
colorama==0.4.6
Django==4.2.8
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
exceptiongroup==1.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
pytest==7.4.3
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
